[Refactor] Move pipeline primitive contracts to ExecPrimitive
