+++ be/src/exec/pipeline/adaptive/collect_stats_context.cpp
@@ -18,8 +18,8 @@
 
 #include "column/chunk.h"
 #include "common/statusor.h"
-#include "exec/pipeline/adaptive/adaptive_dop_param.h"
 #include "exec/pipeline/adaptive/utils.h"
+#include "exec/pipeline/primitives/adaptive_dop_param.h"
 #include "exec/pipeline/primitives/event.h"
 #include "runtime/runtime_state.h"
