+++ be/src/exec/scan_node.h
@@ -44,7 +44,7 @@
 #include "column/column_access_path.h"
 #include "common/runtime_profile.h"
 #include "exec/exec_node.h"
-#include "exec/pipeline/scan/morsel_queue_factory_base.h"
+#include "exec/pipeline/primitives/morsel_queue_factory.h"
 #include "gen_cpp/InternalService_types.h"
 
 namespace starrocks {
