+++ build-support/runtime_state_header_include_allowlist.txt
@@ -1,7 +1,6 @@
 # Headers that still require a complete RuntimeState definition.
 src/exec/join/join_key_constructor.h
 src/exec/pipeline/exchange/local_exchange.h
-src/exec/pipeline/fragment_context.h
 src/exec/pipeline/group_execution/execution_group.h
 src/exec/union_node.h
 src/exprs/agg/array_agg.h
